ADC_Python_Reader/requirements.txt
polars
pyserial
loguru
easygui
hvplot


